Created TextField and Login Components and restructured directory hierarchy to conform to css import protocol
