+++ client/Main.css
@@ -1,1 +1,5 @@
 /* CSS declarations go here */
+
+body {
+    background-color: whitesmoke;
+}
