+++ imports/components/IndexComponent.js
@@ -1,13 +1,15 @@
 import React, {Component} from 'react';
 import {Route, BrowserRouter, Switch} from 'react-router-dom';
 
+import Login from './LoginComponent';
+
 export default class Index extends Component {
 
     render() {
         return (
             <BrowserRouter>
                 <Switch>
-                    <Route exact path='/' component={LayoutComponent} />
+                    <Route exact path='/' component={Login} />
                 </Switch>
             </BrowserRouter>
         );
