+++ imports/components/LoginComponent.js
@@ -1,0 +1,29 @@
+import React, {Component} from 'react';
+import {Redirect} from 'react-router-dom';
+
+import TextField from './TextFieldComponent';
+
+import '../ui/css/LoginStyleSheet.css';
+
+export default class Login extends Component {
+    constructor() {
+        super();
+    }
+
+    render() {
+        return (
+          <form onSubmit={this.login}>
+              <TextField id='email' type='email' placeholder='email'/>
+              <TextField id='password' type='password' placeholder='password'/>
+              <button>Sign Up!</button>
+              <input type='submit' value='Sign In'/>
+          </form>
+        );
+    }
+
+    login = () => {
+        email = document.getElementById('email').value;
+        password = document.getElementById('password').value;
+    }
+
+}
