+++ imports/ui/css/LoginStyleSheet.css
@@ -1,0 +1,65 @@
+@import url(http://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,700italic,300,400,700);
+
+form {
+    /*width: 100%;*/
+    height: 180px;
+    float: right;
+    background-color: #063651;
+    border-radius: 4px;
+}
+
+input[type=submit] {
+    border: 1px solid #D4541B;
+    float: right;
+    width: 123px;
+    height: 35px;
+    background: #D4541B;
+    cursor: pointer;
+    border-radius: 5px;
+    color: #063651/*#A9A9A9*/;
+    font-family: 'Open Sans', sans-serif;
+    font-size: 18px;
+    font-weight: 400;
+    margin-top: 8px;
+    margin-right: 16px;
+}
+
+input[type=submit]:hover {
+    opacity: 0.8;
+}
+
+input[type=submit]:active {
+    opacity: 0.6;
+}
+
+input[type=submit]:focus {
+    outline: none;
+}
+
+button {
+    border: 1px solid #A9A9A9;
+    float: left;
+    width: 123px;
+    height: 35px;
+    background: #A9A9A9;
+    cursor: pointer;
+    border-radius: 5px;
+    color: #063651/*#A9A9A9*/;
+    font-family: 'Open Sans', sans-serif;
+    font-size: 18px;
+    font-weight: 400;
+    margin-top: 8px;
+    margin-left: 16px;
+}
+
+button:hover {
+    opacity: 0.8;
+}
+
+button:active {
+    opacity: 0.6;
+}
+
+button:focus {
+    outline: none;
+}
