+++ imports/ui/css/TextFieldStyleSheet.css
@@ -11,8 +11,9 @@ input[type=text] {
     font-size: 18px;
     font-weight: 400;
     padding: 6px;
-    margin-top: 16px;
+    margin-top: 8px;
     margin-right: 16px;
+    margin-left: 16px;
 }
 
 input[type=password] {
@@ -26,11 +27,34 @@ input[type=password] {
     font-size: 18px;
     font-weight: 400;
     padding: 6px;
-    margin-top: 16px;
+    margin-top: 8px;
     margin-right: 16px;
+    margin-left: 16px;
+}
+
+
+input[type=email] {
+    width: 246px;
+    height: 35px;
+    background: RGBA(212, 84, 27, 0.1);
+    border: 1px solid RGBA(212, 84, 27, 1.00);
+    border-radius: 5px;
+    color: #A9A9A9;
+    font-family: 'Open Sans', sans-serif;
+    font-size: 18px;
+    font-weight: 400;
+    padding: 6px;
+    margin-top: 8px;
+    margin-right: 16px;
+    margin-left: 16px;
 }
 
 input[type=text]:focus {
+    outline: none;
+    border: 1px solid rgba(212, 84, 27, 0.8);
+}
+
+input[type=email]:focus {
     outline: none;
     border: 1px solid rgba(212, 84, 27, 0.8);
 }
